ALAPKÉSZLET? KAMION

+++ src/main/java/com/raktar3/service/ProductService.java
@@ -11,17 +11,14 @@ import org.springframework.stereotype.Service;
 import com.raktar3.entities.Company;
 import com.raktar3.entities.Employe;
 import com.raktar3.entities.Product;
-import com.raktar3.entities.Stock;
 import com.raktar3.repository.CompanyRepository;
-import com.raktar3.repository.EmployeRepository;
 import com.raktar3.repository.ProductRepository;
-import com.raktar3.repository.StockRepository;
 
 @Service
 public class ProductService {
 
 	@Autowired
-	StockRepository stockRepo;
+	StockService stockService;
 	
 	@Autowired
 	ProductRepository productRepo;
@@ -30,10 +27,10 @@ public class ProductService {
 	CompanyRepository companyRepo;
 	
 	@Autowired
-	EmployeRepository employeRepo;
+	EmployeService employeService;
 	
 	
-	public Optional<Product> findById(Integer id) {
+	public Product findById(int id) {
 		return productRepo.findById(id);
 	}
 	
@@ -49,8 +46,11 @@ public class ProductService {
 		productRepo.save(p);
 	}
 	
-	public void deleteProduct(Product p) {
-		productRepo.delete(p);
+	public boolean deleteProduct(Product p) {
+		if (!stockService.vaneProduct(p.getId())) {
+			productRepo.delete(p);
+			return true;
+		} else return false;
 	}
 	
 	public List<Product> findAll(){
@@ -58,17 +58,17 @@ public class ProductService {
 	}
 	
 	public Employe empNameSearch(String name) {
-		return employeRepo.findByNameIgnoreCase(name);
+		return employeService.findByNameIgnoreCase(name);
 	}
 	
 	
 	public List<Employe> findAllEmploye(){
-		return employeRepo.findAll();
+		return employeService.findAllEmploye();
 	}
 	
 	public boolean addUserToDb(Employe emp) {
 		if (empNameSearch(emp.getName())==null) {
-			employeRepo.save(emp);
+			employeService.addUserToDb(emp);
 			return true;
 		} else return false;
 	}
@@ -81,48 +81,54 @@ public class ProductService {
 	public List<Company> findAllCompany(){
 		return companyRepo.findAll();
 	}
+
+	public void deleteAll() {
+		productRepo.deleteAll();
+		
+	}
 	
 	
 	
 	@PostConstruct
 	public void init() {
+//		
+//		Product p = new Product("Aqua Via");productRepo.save(p);
+//		Product p1 = new Product("Brill");productRepo.save(p1);
+//		Product p2 = new Product("Milotai");productRepo.save(p2);
+//		Product p3 = new Product("via-ÜRES BALLON ");productRepo.save(p3);
+//		Product p4 = new Product("brill-ÜRES BALLON");productRepo.save(p4);
+//		Product p5 = new Product("milotai-ÜRES BALLON");productRepo.save(p5);
+//		
+//
+//		Employe e = new Employe("ALAPKÉSZLET");
+//		Employe e1 = new Employe("KAMION");
+//		Employe e2 = new Employe("Vencel");
+//		Employe e3 = new Employe("Sanyi");
+//		
+//		
 		
-		Product p = new Product("Aqua Via");productRepo.save(p);
-		Product p1 = new Product("Brill");productRepo.save(p1);
-		Product p2 = new Product("Milotai");productRepo.save(p2);
-		Product p3 = new Product("via-ÜRES BALLON ");productRepo.save(p3);
-		Product p4 = new Product("brill-ÜRES BALLON");productRepo.save(p4);
-		Product p5 = new Product("milotai-ÜRES BALLON");productRepo.save(p5);
-		
-
-		Employe e = new Employe("KAMION");
-		Employe e1 = new Employe("ALAPKÉSZLET");
-		Employe e2 = new Employe("Vencel");
-		Employe e3 = new Employe("Sanyi");
-		
-		
-		employeRepo.save(e);
-		employeRepo.save(e1);
-		employeRepo.save(e2);
-		employeRepo.save(e3);
-
-		
- 		Company comp = new Company("Bumet","Debrecen","Határ út","H,S,P",e);
-		Company comp1 = new Company("Manz","Debrecen","Határ út","H",e2);		
-		companyRepo.save(comp);
-		companyRepo.save(comp1);
-		
+//		employeService.addUserToDb(e);
+//		employeService.addUserToDb(e1);
+//
+//		
+// 		Company comp = new Company("Bumet","Debrecen","Határ út","H,S,P",e);
+//		Company comp1 = new Company("Manz","Debrecen","Határ út","H",e2);		
+//		companyRepo.save(comp);
+//		companyRepo.save(comp1);
+//		
 //		Stock s = new Stock(50,"2019-02-08",true,e,p,"nincs komment");
 //		Stock s2 = new Stock(120,"2019-02-07",true,e2,p2,"második");
 //		Stock s3 = new Stock(15,"2019-02-08",false,e,p,"eladas");
 //		Stock s4 = new Stock(70,"2019-02-07",false,e2,p2,"eladas");
 //		
-//		stockRepo.save(s);
-//		stockRepo.save(s2);
-//		stockRepo.save(s3);
-//		stockRepo.save(s4);
+//		stockService.addIncoming(s);
+//		stockService.addIncoming(s2);
+//		stockService.addIncoming(s3);
+//		stockService.addIncoming(s4);
 //		
 	}
+
+
